refactor(server): ルートをクラスベースのControllerに変換
DIコンテナの直接引き渡しをやめ、inversifyによるコンストラクタ
インジェクションを使用するControllerクラスに変換。

- 10個のControllerクラスを作成（ImageController, LabelController等）
- container.tsにControllerバインディングを追加
- AppContainerにController取得メソッドを追加
- 旧ルートファイル（10ファイル）を削除
