+++ .ai-agent/steering/idea.md
@@ -16,7 +16,6 @@
 | 内容 | 温度感 |
 |------|--------|
 | 設定ファイルのパスを指定可能にする | 高 |
-| DI注入をクラス化して行い、コンテナの直接の引き渡しを止める | 高 |
 | タイトル自動生成。ID使うのをやめ、どういう画像なのか判別しやすくする | 中 |
 | 閲覧・編集モード分離 | 中 |
 | encodeURIComponent の使用をやめ、URLビルダを利用する | 中 |
