+++ packages/server/src/infra/di/types.ts
@@ -22,4 +22,16 @@ export const TYPES = {
   EmbeddingService: Symbol.for('EmbeddingService'),
   EmbeddingRepository: Symbol.for('EmbeddingRepository'),
   CaptionService: Symbol.for('CaptionService'),
+
+  // Controllers
+  ImageController: Symbol.for('ImageController'),
+  ImageAttributeController: Symbol.for('ImageAttributeController'),
+  LabelController: Symbol.for('LabelController'),
+  CollectionController: Symbol.for('CollectionController'),
+  ArchiveController: Symbol.for('ArchiveController'),
+  ViewHistoryController: Symbol.for('ViewHistoryController'),
+  RecommendationController: Symbol.for('RecommendationController'),
+  RecommendationConversionController: Symbol.for('RecommendationConversionController'),
+  StatsController: Symbol.for('StatsController'),
+  SearchController: Symbol.for('SearchController'),
 } as const;
