+++ packages/server/src/infra/http/controllers/index.ts
@@ -1,0 +1,10 @@
+export { ArchiveController } from './archive-controller.js';
+export { CollectionController } from './collection-controller.js';
+export { ImageAttributeController } from './image-attribute-controller.js';
+export { ImageController } from './image-controller.js';
+export { LabelController } from './label-controller.js';
+export { RecommendationController } from './recommendation-controller.js';
+export { RecommendationConversionController } from './recommendation-conversion-controller.js';
+export { SearchController } from './search-controller.js';
+export { StatsController } from './stats-controller.js';
+export { ViewHistoryController } from './view-history-controller.js';
