+++ packages/server/src/infra/http/routes/index.ts
@@ -1,14 +1,4 @@
-import { archiveRoutes } from '@/infra/http/routes/archives.js';
-import { collectionRoutes } from '@/infra/http/routes/collections.js';
 import { healthRoutes } from '@/infra/http/routes/health.js';
-import { imageAttributeRoutes } from '@/infra/http/routes/image-attributes.js';
-import { imageRoutes } from '@/infra/http/routes/images.js';
-import { labelRoutes } from '@/infra/http/routes/labels.js';
-import { recommendationConversionRoutes } from '@/infra/http/routes/recommendation-conversions.js';
-import { recommendationRoutes } from '@/infra/http/routes/recommendations.js';
-import { searchRoutes } from '@/infra/http/routes/search.js';
-import { statsRoutes } from '@/infra/http/routes/stats.js';
-import { viewHistoryRoutes } from '@/infra/http/routes/view-history.js';
 import type { AppContainer } from '@/infra/di/index.js';
 import type { FastifyInstance } from 'fastify';
 
@@ -16,33 +6,15 @@ export function registerRoutes(app: FastifyInstance, container: AppContainer): v
   // Health check (no prefix, no container dependency)
   healthRoutes(app);
 
-  // Image routes
-  imageRoutes(app, container);
-
-  // Image attribute routes
-  imageAttributeRoutes(app, container);
-
-  // Label routes
-  labelRoutes(app, container);
-
-  // Collection routes
-  collectionRoutes(app, container);
-
-  // Archive routes
-  archiveRoutes(app, container);
-
-  // View history routes
-  viewHistoryRoutes(app, container);
-
-  // Recommendation routes
-  recommendationRoutes(app, container);
-
-  // Recommendation conversion routes
-  recommendationConversionRoutes(app, container);
-
-  // Stats routes
-  statsRoutes(app, container);
-
-  // Search routes
-  searchRoutes(app, container);
+  // Register controllers with inversify DI
+  container.getImageController().registerRoutes(app);
+  container.getImageAttributeController().registerRoutes(app);
+  container.getLabelController().registerRoutes(app);
+  container.getCollectionController().registerRoutes(app);
+  container.getArchiveController().registerRoutes(app);
+  container.getViewHistoryController().registerRoutes(app);
+  container.getRecommendationController().registerRoutes(app);
+  container.getRecommendationConversionController().registerRoutes(app);
+  container.getStatsController().registerRoutes(app);
+  container.getSearchController().registerRoutes(app);
 }
